// src/types/nationalities.ts
export type NationalitiesByName = Record<string, CountryResult[]>;

export type COUNTRY_ID =
  | 'FR'
  | 'IT'
  | 'ES'
  | 'TR'
  | 'DE'
  | 'PL'
  | 'GB'
  | 'RU'
  | 'CZ'
  | 'PT'
  | 'NL'
  | 'BE'
  | 'MA'
  | 'PH'
  | 'US'
  | 'RO'
  | 'DZ'
  | 'NG'
  | 'CH'
  | 'HU'
  | 'TH'
  | 'SE'
  | 'ID'
  | 'IN'
  | 'UA'
  | 'MY'
  | 'TN'
  | 'SA'
  | 'GR'
  | 'CI'
  | 'AT'
  | 'ZA'
  | 'KR'
  | 'CN'
  | 'RS'
  | 'JP'
  | 'EG'
  | 'SK'
  | 'SN'
  | 'DK'
  | 'FI'
  | 'CM'
  | 'IR'
  | 'AR'
  | 'CA'
  | 'SG'
  | 'PK'
  | 'GH'
  | 'LB'
  | 'IE'
  | 'AO'
  | 'NO'
  | 'BY'
  | 'BR'
  | 'MX'
  | 'CO'
  | 'KE'
  | 'CL'
  | 'KW'
  | 'AL'
  | 'VE'
  | 'RE'
  | 'BA'
  | 'IL'
  | 'TW'
  | 'SI'
  | 'KZ'
  | 'PE'
  | 'AZ'
  | 'AE'
  | 'CY'
  | 'LT'
  | 'DO'
  | 'JO'
  | 'MD'
  | 'BJ'
  | 'BG'
  | 'CD'
  | 'HR'
  | 'LV'
  | 'HK'
  | 'MZ'
  | 'AU'
  | 'LU'
  | 'UG'
  | 'ML'
  | 'BF'
  | 'MU'
  | 'OM'
  | 'TG'
  | 'QA'
  | 'MK'
  | 'MG'
  | 'VN'
  | 'GA'
  | 'EE'
  | 'IQ'
  | 'MT'
  | 'BH'
  | 'TZ'
  | 'EC'
  | 'GE'
  | 'AM'
  | 'SD'
  | 'ET'
  | 'MM'
  | 'ME'
  | 'SY'
  | 'UZ'
  | 'ZW'
  | 'DJ'
  | 'LK'
  | 'BD'
  | 'SH'
  | 'BW'
  | 'CV'
  | 'BO'
  | 'YE'
  | 'RW'
  | 'IS'
  | 'LY'
  | 'NE'
  | 'AD'
  | 'GM'
  | 'CG'
  | 'ZM'
  | 'AF'
  | 'NA'
  | 'MR'
  | 'UY'
  | 'CR'
  | 'KG'
  | 'PA'
  | 'NP'
  | 'GN'
  | 'GP'
  | 'GQ'
  | 'MQ'
  | 'SC'
  | 'CU'
  | 'NZ'
  | 'GT'
  | 'MC'
  | 'MV'
  | 'MW'
  | 'SV'
  | 'GF'
  | 'SL'
  | 'LR'
  | 'NI'
  | 'KH'
  | 'TJ'
  | 'PR'
  | 'BI'
  | 'PY'
  | 'KM'
  | 'LS'
  | 'SO'
  | 'HN'
  | 'TD'
  | 'GI'
  | 'BN'
  | 'SM'
  | 'JE'
  | 'SZ'
  | 'TM'
  | 'NC'
  | 'MN'
  | 'LI'
  | 'GL'
  | 'BT'
  | 'PF'
  | 'ST'
  | 'YT'
  | 'FO'
  | 'JM'
  | 'HT'
  | 'BS'
  | 'GW'
  | 'FJ'
  | 'GG'
  | 'LA'
  | 'TL'
  | 'TT'
  | 'AG'
  | 'EH'
  | 'CF'
  | 'PW'
  | 'IM'
  | 'AW'
  | 'BZ'
  | 'AS'
  | 'AI'
  | 'MP'
  | 'BM'
  | 'ER'
  | 'SR'
  | 'VG'
  | 'VI'
  | 'GY'
  | 'BB'
  | 'MH'
  | 'MO'
  | 'PG'
  | 'KY'
  | 'KI'
  | 'AN'
  | 'CK'
  | 'GU'
  | 'WF'
  | 'PS'
  | 'TC'
  | 'VU'
  | 'LC'
  | 'TO'
  | 'CC'
  | 'WS'
  | 'GD'
  | 'VA'
  | 'IO'
  | 'FM'
  | 'NU'
  | 'BL'
  | 'KN'
  | 'PM'
  | 'FK'
  | 'AQ'
  | 'PN'
  | 'MF'
  | 'TV'
  | 'VC'
  | 'DM'
  | 'SB'
  | 'TK'
  | 'NF'
  | 'CX'
  | 'NR'
  | 'SJ'
  | 'MS'
  | 'KP'
  | 'GS'
  | 'BV';

export type CountryResult = {
  country_id: COUNTRY_ID;
  probability: number;
};

export type NameResult = {
  count: number;
  name: string;
  country: CountryResult[];
};

export type NationalityResponse = NameResult[];